[CMP-9092](https://youtrack.jetbrains.com/issue/CMP-9092) Merge Jetpack
Compose 1.10.0-beta02

| GroupId | ReleaseVersion | ReleaseSHA | ReleaseBuildId | ReleaseDate |
| --- | --- | --- | --- | --- |
| androidx.compose.animation | 1.10.0-beta02 |
d1c84233fc2372352b838d165d256581ff37ada9 | 14361795 | 11/5/2025 |
| androidx.compose.foundation | 1.10.0-beta02 |
d1c84233fc2372352b838d165d256581ff37ada9 | 14361795 | 11/5/2025 |
| androidx.compose.material | 1.10.0-beta02 |
d1c84233fc2372352b838d165d256581ff37ada9 | 14361795 | 11/5/2025 |
| androidx.compose.runtime | 1.10.0-beta02 |
d1c84233fc2372352b838d165d256581ff37ada9 | 14361795 | 11/5/2025 |
| androidx.compose.ui | 1.10.0-beta02 |
d1c84233fc2372352b838d165d256581ff37ada9 | 14361795 | 11/5/2025 |

## Release Notes
N/A

+++ compose/runtime/runtime/src/jvmMain/kotlin/androidx/compose/runtime/reflect/ComposableMethod.jvm.kt
@@ -77,7 +77,7 @@ private fun Method.getComposableInfo(): ComposableInfo {
  */
 private fun Class<*>.getDefaultValue(): Any? =
     when (name) {
-        "int" -> 0.toInt()
+        "int" -> 0
         "short" -> 0.toShort()
         "byte" -> 0.toByte()
         "long" -> 0.toLong()
